+++ libs/components/core/src/lib/modules/resize-observer/resize-observer-media-query.service.ts
@@ -22,12 +22,12 @@ export class SkyResizeObserverMediaQueryService implements OnDestroy {
     return this._currentBreakpoint;
   }
 
-  private _breakpoints: {
+  #breakpoints: {
     check: (width: number) => boolean;
     name: SkyMediaBreakpoints;
   }[] = [
     {
-      check: (width: number) => width <= 767,
+      check: (width: number) => width > 0 && width <= 767,
       name: SkyMediaBreakpoints.xs,
     },
     {
@@ -43,26 +43,29 @@ export class SkyResizeObserverMediaQueryService implements OnDestroy {
       name: SkyMediaBreakpoints.lg,
     },
   ];
-  private _currentBreakpointObservable = new ReplaySubject<
+
+  #currentBreakpointObservable = new ReplaySubject<
     SkyMediaBreakpoints | undefined
   >(1);
+
   private _currentBreakpoint: SkyMediaBreakpoints;
-  private _resizeSubscription: Subscription;
-  private _stopListening = new Subject<void>();
-  private _target?: ElementRef;
+
+  #stopListening = new Subject<void>();
+
+  #target: ElementRef;
 
   constructor(private resizeObserverService: SkyResizeObserverService) {
-    this._stopListening.subscribe(() => {
-      this._target = undefined;
+    this.#stopListening.subscribe(() => {
+      this.#target = undefined;
       this.updateBreakpoint(undefined);
     });
   }
 
   public ngOnDestroy(): void {
-    this._stopListening.next();
+    this.#stopListening.next();
     this._currentBreakpoint = undefined;
-    this._stopListening.complete();
-    this._currentBreakpointObservable.complete();
+    this.#stopListening.complete();
+    this.#currentBreakpointObservable.complete();
   }
 
   /**
@@ -77,25 +80,25 @@ export class SkyResizeObserverMediaQueryService implements OnDestroy {
    * time. Any previous subscriptions will be unsubscribed when a new element is observed.
    */
   public observe(element: ElementRef): SkyResizeObserverMediaQueryService {
-    if (this._target) {
-      if (this._target === element) {
+    if (this.#target) {
+      if (this.#target === element) {
         return this;
       }
-      this._stopListening.next();
+
+      this.#stopListening.next();
     }
-    this._target = element;
-    const width = (element.nativeElement as HTMLElement).offsetWidth;
-    if (width) {
-      const breakpoint = this.checkBreakpoint(width);
-      this.updateBreakpoint(breakpoint);
-    }
-    this._resizeSubscription = this.resizeObserverService
+
+    this.#target = element;
+
+    this.checkWidth(element);
+
+    this.resizeObserverService
       .observe(element)
-      .pipe(takeUntil(this._stopListening))
+      .pipe(takeUntil(this.#stopListening))
       .subscribe((value) => {
         const breakpoint = this.checkBreakpoint(value.contentRect.width);
         /* istanbul ignore else */
-        if (breakpoint !== this._currentBreakpoint) {
+        if (breakpoint !== this.current) {
           this.updateBreakpoint(breakpoint);
         }
       });
@@ -106,25 +109,42 @@ export class SkyResizeObserverMediaQueryService implements OnDestroy {
    * Stop watching the container element.
    */
   public unobserve(): void {
-    this._stopListening.next();
+    this.#stopListening.next();
   }
 
   /**
    * Subscribes to element size changes that cross breakpoints.
    */
   public subscribe(listener: SkyMediaQueryListener): Subscription {
-    return this._currentBreakpointObservable
-      .pipe(takeUntil(this._stopListening))
-      .subscribe(listener);
+    return this.#currentBreakpointObservable
+      .pipe(takeUntil(this.#stopListening))
+      .subscribe((value) => {
+        listener(value);
+      });
   }
 
   private updateBreakpoint(breakpoint: SkyMediaBreakpoints) {
     this._currentBreakpoint = breakpoint;
-    this._currentBreakpointObservable.next(breakpoint);
+    this.#currentBreakpointObservable.next(breakpoint);
   }
 
   private checkBreakpoint(width: number): SkyMediaBreakpoints | undefined {
-    return this._breakpoints.find((breakpoint) => breakpoint.check(width))
-      ?.name;
+    const breakpoint = this.#breakpoints.find((breakpoint) =>
+      breakpoint.check(width)
+    );
+
+    /* istanbul ignore else */
+    if (breakpoint) {
+      return breakpoint.name;
+    }
+  }
+
+  private checkWidth(element: ElementRef) {
+    const width = (element.nativeElement as HTMLElement).offsetWidth || 0;
+    const breakpoint = this.checkBreakpoint(width);
+    /* istanbul ignore else */
+    if (breakpoint !== this._currentBreakpoint) {
+      this.updateBreakpoint(breakpoint);
+    }
   }
 }