+++ libs/components/core/src/lib/modules/resize-observer/resize-observer.service.ts
@@ -16,19 +16,18 @@ type ResizeObserverTracking = {
   providedIn: 'any',
 })
 export class SkyResizeObserverService implements OnDestroy {
-  private _resizeObserver: ResizeObserver;
-  private _tracking: ResizeObserverTracking[] = [];
+  #resizeObserver: ResizeObserver;
+
+  #tracking: ResizeObserverTracking[] = [];
 
   constructor(private zone: NgZone) {
-    this._resizeObserver = new ResizeObserver(
-      (entries: ResizeObserverEntry[]) => {
-        entries.forEach((entry) => this.callback(entry));
-      }
-    );
+    this.#resizeObserver = new ResizeObserver((entries) => {
+      entries.forEach((entry) => this.callback(entry));
+    });
   }
 
   public ngOnDestroy(): void {
-    this._resizeObserver.disconnect();
+    this.#resizeObserver.disconnect();
   }
 
   /**
@@ -39,43 +38,51 @@ export class SkyResizeObserverService implements OnDestroy {
   }
 
   private observeAndTrack(element: ElementRef): ResizeObserverTracking {
-    const checkTracking = this._tracking.findIndex((value) => {
+    const checkTracking = this.#tracking.findIndex((value) => {
       return !value.subject.closed && value.element === element.nativeElement;
     });
+
     if (checkTracking === -1) {
-      this._resizeObserver.observe(element.nativeElement);
+      this.#resizeObserver.observe(element.nativeElement);
     }
+
     const subject = new Subject<ResizeObserverEntry>();
     const subjectObservable = subject.pipe(
       finalize(() => {
         // Are there any other tracking entries still watching this element?
-        const checkTracking = this._tracking.findIndex((value) => {
+        const checkTracking = this.#tracking.findIndex((value) => {
           return (
             value.subject !== subject &&
             !value.subject.closed &&
             value.element === element.nativeElement
           );
         });
+
         if (checkTracking === -1) {
-          this._resizeObserver.unobserve(element.nativeElement);
+          this.#resizeObserver.unobserve(element.nativeElement);
         }
       })
     );
+
     const tracking = {
       element: element.nativeElement,
       subject,
       subjectObservable,
     };
-    this._tracking.push(tracking);
+
+    this.#tracking.push(tracking);
+
     return tracking;
   }
 
-  private callback(entry: ResizeObserverEntry) {
-    this._tracking
+  private callback(entry: ResizeObserverEntry): void {
+    this.#tracking
       .filter((value) => !(value.subject.closed || value.subject.isStopped))
       .forEach((value) => {
         /* istanbul ignore else */
         if (value.element === entry.target) {
+          // Execute the callback within NgZone because Angular does not "monkey patch"
+          // ResizeObserver like it does for other features in the DOM.
           this.zone.run(() => {
             value.subject.next(entry);
           });