+++ libs/components/lists/src/lib/modules/repeater/repeater.service.ts
@@ -2,6 +2,7 @@ import { EventEmitter, Injectable, OnDestroy } from '@angular/core';
 
 import { BehaviorSubject } from 'rxjs';
 
+import { SkyRepeaterItemRolesType } from './repeater-item-roles.type';
 import { SkyRepeaterItemComponent } from './repeater-item.component';
 
 /**
@@ -23,6 +24,12 @@ export class SkyRepeaterService implements OnDestroy {
 
   public items: SkyRepeaterItemComponent[] = [];
 
+  public readonly itemRole = new BehaviorSubject<SkyRepeaterItemRolesType>({
+    content: undefined,
+    item: undefined,
+    title: undefined,
+  });
+
   public orderChange = new BehaviorSubject<void>(undefined);
 
   public repeaterGroupId: number;
@@ -36,7 +43,7 @@ export class SkyRepeaterService implements OnDestroy {
   public activateItem(item: SkyRepeaterItemComponent): void {
     if (this.enableActiveState) {
       /* istanbul ignore else */
-      if (item) {
+      if (item && !item.disabled) {
         const index = this.items.findIndex((i) => i === item);
         this.activeItemIndexChange.next(index);
         this.activeItemChange.next(item);
@@ -51,7 +58,7 @@ export class SkyRepeaterService implements OnDestroy {
         this.activeItemChange.next(undefined);
       } else {
         const activeItem = this.items[index];
-        if (activeItem) {
+        if (activeItem && !activeItem.disabled) {
           this.activeItemChange.next(activeItem);
         }
       }