+++ libs/components/modals/src/lib/modules/modal/modal.component.ts
@@ -15,7 +15,6 @@ import {
   SkyCoreAdapterService,
   SkyDockLocation,
   SkyDockService,
-  SkyMediaQueryService,
   SkyResizeObserverMediaQueryService,
 } from '@skyux/core';
 
@@ -38,14 +37,7 @@ let skyModalUniqueIdentifier = 0;
   templateUrl: './modal.component.html',
   styleUrls: ['./modal.component.scss'],
   animations: [skyAnimationModalState],
-  providers: [
-    SkyModalComponentAdapterService,
-    SkyDockService,
-    {
-      provide: SkyMediaQueryService,
-      useClass: SkyResizeObserverMediaQueryService,
-    },
-  ],
+  providers: [SkyModalComponentAdapterService, SkyDockService],
 })
 export class SkyModalComponent implements AfterViewInit, OnDestroy {
   @HostBinding('class')
@@ -130,7 +122,6 @@ export class SkyModalComponent implements AfterViewInit, OnDestroy {
     private coreAdapter: SkyCoreAdapterService,
     @Host() private dockService: SkyDockService,
     @Optional()
-    @Host()
     private mediaQueryService?: SkyResizeObserverMediaQueryService
   ) {}